code : BOJ 1620 s4 나는야 포켓몬 마스터 이다솜
